backend/requirements.txt
fastapi
'uvicorn[standard]'
pillow
python-dotenv
langchain
langchain_google_genai
pinecone-client
torch
git+https://github.com/openai/CLIP.git

